perf(jit): static-action fast path drops cold compile 3h+ → 21min
dispatch_action's 46-handler ``jax.lax.switch`` inlines every handler
into HLO at trace time so XLA's conditional-canonicalizer pass walked
a quadratic graph and stalled on cold compile (Mac CPU 3h+, H100 3h+,
both with zero seed completions).

For callers passing ``action`` as a Python int (the byte-parity
validator's fixed action=0), the target handler is fully determined
at trace time.  Pick it host-side via the existing
``_ACTION_TO_HANDLER_IDX`` / ``_SLOT_TO_COMPACT`` / ``_SLOT_TO_DIR_IDX``
tables and call the chosen ``_COMPACT_HANDLERS[i]`` directly — only
that single handler enters HLO.

* New ``_dispatch_validator_body`` / ``_dispatch_jit_validator`` in
  ``env.py`` (jit'd with ``static_argnums=(1,)``).
* ``_step_impl`` routes through the fast path when ``action`` is a
  Python int, falls through to ``_dispatch_jit`` otherwise.
* Validator now passes ``action`` (int) instead of ``jnp.int32(action)``.

Verified Mac CPU 10/10 PASS preserved in eager mode and 2/2 PASS in
JIT mode at 21:06 wallclock (single cold compile).  Previous bare
validator JIT compile never terminated in 3h+.

Assumption: pending_action prompt state stays empty for action=0 on
Main Dlvl 1 (validator's path).  The generic ``_dispatch_jit`` path
remains the default for non-static action callers.

+++ tests/test_nle_byte_parity.py
@@ -220,8 +220,12 @@ def run_validator(
 
         # Step Nethax
         try:
+            # Pass ``action`` as a Python int so NethaxEnv routes through
+            # the static-action fast path (env.py::_dispatch_jit_validator),
+            # avoiding the 46-branch lax.switch in dispatch_action that
+            # otherwise stalls JIT cold-compile.
             nax_state, nax_obs, _, nax_done, _ = nax_env.step(
-                nax_state, jnp.int32(action), jax.random.PRNGKey(seed + step_idx)
+                nax_state, action, jax.random.PRNGKey(seed + step_idx)
             )
         except Exception as e:
             print(f"[abort] Nethax step {step_idx} failed: {e}")
